cmd/faces: add -s to match mailfs -s (and Mail -n)
Today I was running two copies of mailfs, one with -s domain1 one
with -s domain2, and corresponding instances of Mail, and noticed I
couldn't similarly have two copies of faces.

This change allows one to specify what mailfs faces should attach to.
(It's not possible to run two instances of faces, because plumb
messages for seemail don't contain the originating mailfs server
name, so it's a bit of a mess; but at least with this change I can
start faces when the defailt "mail" server name is not used.)